+++ test/pipeline_test.rb
@@ -58,6 +58,14 @@ class PipelineTest < Minitest::Test
     assert_in_delta (-0.3130), output[-1][-1][-1]
   end
 
+  def test_embedding
+    sentences = ["This is an example sentence", "Each sentence is converted"]
+    embed = Informers.pipeline("embedding")
+    embeddings = embed.(sentences)
+    assert_elements_in_delta [0.067657, 0.063496, 0.048713], embeddings[0][..2]
+    assert_elements_in_delta [0.086439, 0.10276, 0.0053946], embeddings[1][..2]
+  end
+
   def test_rerank
     ranker = Informers.pipeline("rerank")
     result = ranker.("Who created Ruby?", ["Matz created Ruby", "Another doc"])